+++ datafiles/update_node_project.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+cd /skale/skale-node
+sudo git pull https://$GITHUB_TOKEN\@github.com/skalenetwork/skale-node.git